refactor(core): allow configure options and load source
The current implementation ignores the options object if it has a `source` or `sources` attribute.
This will cause an incorrect behavior: when user want to reconfigure some options and load a new source (at the same time), this will not happen, because the configure method will ignore the options in favor of loading the source.
This change will ensure that the two actions (load/configure) happen.